shop_api/views/store.py
from rest_framework import generics

from shop_api.models import Store
from shop_api.serializers import StoreSerializer


class StoreList(generics.ListCreateAPIView):
    queryset = Store.objects.all()
    serializer_class = StoreSerializer


class StoreDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Store.objects.all()
    serializer_class = StoreSerializer
